+++ src/data/data.js
@@ -1,0 +1,31 @@
+
+export const filtersList = [
+  {
+    name: `all`,
+    count: 15
+  },
+  {
+    name: `overdue`,
+    count: 0
+  },
+  {
+    name: `today`,
+    count: 0
+  },
+  {
+    name: `favorites`,
+    count: 7
+  },
+  {
+    name: `repeating`,
+    count: 2
+  },
+  {
+    name: `tags`,
+    count: 6
+  },
+  {
+    name: `archive`,
+    count: 115
+  }
+];
